Refactor Postman
