+++ pkg/postman/trisa.go
@@ -1,0 +1,188 @@
+/*
+Postman providers helper functionality for managing TRISA and TRP transfers and
+sorting them and storing them in the database. This package is intended to unify the
+functionality across the TRISA node, the TRP node, and the Web API/UI.
+
+On every single travel rule transaction, no matter if it's TRISA or TRP, no matter
+if it's sent from the node or received into the node, whether or not it's a new
+transaction or an update to an old transaction the following things must happen:
+
+1. The message(s) must be validated
+2. The transfer packet must be associated with a transaction
+3. The transaction status must be updated, and potentially other parts of the transaction
+4. The counterparty must be identified
+5. Error envelopes have to be handled correctly
+6. The keys for the envelope must be loaded for decryption
+7. Sealed envelopes need to be decrypted
+8. HMAC signatures need to be checked
+9. The outgoing envelope must be resealed with internal keys
+10. The envelopes and all changes must be saved to the database
+11. The audit log must be updated
+*/
+package postman
+
+import (
+	"database/sql"
+	"fmt"
+
+	"github.com/google/uuid"
+	"github.com/rs/zerolog"
+
+	"github.com/trisacrypto/envoy/pkg/trisa/peers"
+
+	api "github.com/trisacrypto/trisa/pkg/trisa/api/v1beta1"
+	"github.com/trisacrypto/trisa/pkg/trisa/envelope"
+)
+
+type TRISAPacket struct {
+	Packet
+	Peer     peers.Peer  // The remote peer the transfer is being conducted with
+	PeerInfo *peers.Info // The peer info for finding the counterparty
+}
+
+func SendTRISA(payload *api.Payload, envelopeID uuid.UUID, transferState api.TransferState, log zerolog.Logger) (packet *TRISAPacket, err error) {
+	packet = &TRISAPacket{
+		Packet: Packet{
+			In:      &Incoming{},
+			Out:     &Outgoing{},
+			Log:     log,
+			Request: DirectionOutgoing,
+			Reply:   DirectionIncoming,
+		},
+	}
+
+	// Add parent to submessages
+	packet.In.packet = &packet.Packet
+	packet.Out.packet = &packet.Packet
+	packet.resolver = packet
+
+	opts := []envelope.Option{
+		envelope.WithEnvelopeID(envelopeID.String()),
+		envelope.WithTransferState(transferState),
+	}
+
+	if packet.Out.Envelope, err = envelope.New(payload, opts...); err != nil {
+		return nil, fmt.Errorf("could not create envelope for payload: %w", err)
+	}
+
+	return packet, nil
+}
+
+func SendTRISAReject(reject *api.Error, envelopeID uuid.UUID, log zerolog.Logger) (packet *TRISAPacket, err error) {
+	packet = &TRISAPacket{
+		Packet: Packet{
+			In:      &Incoming{},
+			Out:     &Outgoing{},
+			Log:     log,
+			Request: DirectionOutgoing,
+			Reply:   DirectionIncoming,
+		},
+	}
+
+	// Add parent to submessages
+	packet.In.packet = &packet.Packet
+	packet.Out.packet = &packet.Packet
+	packet.resolver = packet
+
+	// The envelope package should set the correct transfer state based on retry.
+	if packet.Out.Envelope, err = envelope.WrapError(reject, envelope.WithEnvelopeID(envelopeID.String())); err != nil {
+		return nil, fmt.Errorf("could not create rejection envelope: %w", err)
+	}
+
+	return packet, nil
+}
+
+func ReceiveTRISA(in *api.SecureEnvelope, log zerolog.Logger, peer peers.Peer) (packet *TRISAPacket, err error) {
+	packet = &TRISAPacket{
+		Packet: Packet{
+			In:      &Incoming{original: in},
+			Out:     &Outgoing{},
+			Log:     log,
+			Request: DirectionIncoming,
+			Reply:   DirectionOutgoing,
+		},
+		Peer: peer,
+	}
+
+	// Add parent to submessages
+	packet.In.packet = &packet.Packet
+	packet.Out.packet = &packet.Packet
+	packet.resolver = packet
+
+	if packet.In.Envelope, err = envelope.Wrap(packet.In.original); err != nil {
+		return nil, fmt.Errorf("could not wrap incoming secure envelope: %w", err)
+	}
+
+	if packet.PeerInfo, err = peer.Info(); err != nil {
+		return nil, fmt.Errorf("could not identify counterparty in transaction: %w", err)
+	}
+
+	packet.Counterparty = packet.PeerInfo.Model()
+	return packet, nil
+}
+
+// Receive updates the incoming message with the specified secure envelope, e.g. in the
+// case where the outgoing message has been sent and this is the reply that was received
+// from the remote server.
+func (p *TRISAPacket) Receive(in *api.SecureEnvelope) (err error) {
+	p.In = &Incoming{
+		original: in,
+		packet:   &p.Packet,
+	}
+
+	if p.In.Envelope, err = envelope.Wrap(in); err != nil {
+		return fmt.Errorf("could not wrap incoming secure envelope: %w", err)
+	}
+
+	return nil
+}
+
+// Creates a TRISA payload envelope to as the outgoing message in order to send a
+// response back to the remote that initiated the transfer.
+func (p *TRISAPacket) Send(payload *api.Payload, state api.TransferState) (err error) {
+	if p.Out.Envelope, err = p.In.Envelope.Update(payload, envelope.WithTransferState(state)); err != nil {
+		p.Log.Debug().Err(err).Msg("could not prepare outgoing payload")
+		return fmt.Errorf("could not create outgoing payload: %w", err)
+	}
+	return nil
+}
+
+// Reject creates an outgoing message with a TRISA error that contains the specified
+// TRISA rejection error, e.g. in the case where an incoming message is being handled.
+func (p *TRISAPacket) Reject(code api.Error_Code, message string, retry bool) error {
+	reject := &api.Error{
+		Code: code, Message: message, Retry: retry,
+	}
+	return p.Error(reject)
+}
+
+// Creates a TRISA error envelope from the TRISA error message and sets it as the
+// outgoing message (e.g. in the case where an incoming message is being handled).
+func (p *TRISAPacket) Error(reject *api.Error, opts ...envelope.Option) (err error) {
+	if p.Out.Envelope, err = p.In.Envelope.Reject(reject, opts...); err != nil {
+		p.Log.Debug().Err(err).Msg("could not prepare rejection envelope")
+		return fmt.Errorf("could not create rejection envelope: %w", err)
+	}
+	return nil
+}
+
+func (p *TRISAPacket) ResolveCounterparty() (err error) {
+	if p.Counterparty == nil {
+		if p.PeerInfo == nil {
+			if p.Peer == nil {
+				return ErrNoCounterpartyInfo
+			}
+
+			if p.PeerInfo, err = p.Peer.Info(); err != nil {
+				return err
+			}
+		}
+
+		p.Counterparty = p.PeerInfo.Model()
+	}
+	return nil
+}
+
+func (p *TRISAPacket) Remote() sql.NullString {
+	return sql.NullString{Valid: p.PeerInfo.CommonName != "", String: p.PeerInfo.CommonName}
+}
