+++ pkg/web/prepared.go
@@ -103,7 +103,7 @@ func (s *Server) SendPreparedTransaction(c *gin.Context) {
 		err     error
 		in      *api.Prepared
 		payload *trisa.Payload
-		packet  *postman.Packet
+		packet  *postman.TRISAPacket
 		out     *api.Transaction
 	)
 
@@ -133,7 +133,7 @@ func (s *Server) SendPreparedTransaction(c *gin.Context) {
 	log := logger.Tracing(ctx).With().Str("envelope_id", envelopeID.String()).Logger()
 
 	// Create a packet to begin the sending process
-	if packet, err = postman.Send(payload, envelopeID, trisa.TransferStarted, log); err != nil {
+	if packet, err = postman.SendTRISA(payload, envelopeID, trisa.TransferStarted, log); err != nil {
 		c.Error(err)
 		c.JSON(http.StatusInternalServerError, api.Error("could not process send prepared transaction request"))
 		return
